refactoring/add validation in detail information
