+++ src/container/list/components/QuestionList/index.tsx
@@ -2,8 +2,8 @@ import QuestionCard from '../QuestionCard';
 import React from 'react';
 import { QuestionListStyled } from './index.style';
 import { useGetListQuery } from '../../../../store/api';
-import { validateText } from './questionTextValidation';
-import { truncateText } from './questionTextTruncation';
+import { validateText } from '../../../../services/validation/questions/questionTextValidation';
+import { truncateText } from '../../../../services/validation/questions/questionTextTruncation';
 import { getNavigationValue } from '@brojs/cli';
 import { LinkStyled } from './LinkStyled/index.style';
 import Loading from '../../../../components/Loading';
